+++ .github/workflows/lint.env
@@ -1,0 +1,13 @@
+PUBLIC_API_KEY=1234567890ABCDEF
+
+PUBLIC_AUTH_DOMAIN=lint.firebaseapp.com
+
+PUBLIC_PROJECT_ID=lint
+
+PUBLIC_STORAGE_BUCKET=lint.appspot.com
+
+PUBLIC_MESSAGING_SENDER_ID=1122334455
+
+PUBLIC_APP_ID=1:1122334455:web:abc123def456
+
+PUBLIC_MEASUREMENT_ID=A-BCDEF12345
