+++ src/routes/admin/+page.svelte
@@ -1,6 +1,4 @@
 <script>
-
 </script>
 
-<h1> Welcome to admin </h1>
-
+<h1>Welcome to admin</h1>
