+++ vite.config.ts
@@ -2,7 +2,7 @@ import { sveltekit } from '@sveltejs/kit/vite';
 import type { UserConfig } from 'vite';
 
 const config: UserConfig = {
-    plugins: [sveltekit()],
+    plugins: [sveltekit()]
     // if uncommented, resources load as expected
     // server: {
     //     fs: {
